requirements.txt
jupyterlab
pandas
numpy
matplotlib
gym
torch
stable-baselines3[extra]
python-binance
mexc-api-sdk
pyyaml
